Don't raise the state of FragmentManager while destroying.
Bug 34759474

In onDestroy, FragmentActivity was calling FragmentController's
doReallyStop(), which moved the state to STOPPED. This can
raise the state and call onCreate() on Fragments that weren't
previously raised to that state.

The problem is that the initial state of the FragmentActivity
was not set to "stopped," so it thought that it had to destroy
the FragmentController when it didn't.

Test: gradlew connectedCheck in Fragments on sailfish

Change-Id: I91e8a15b88380b8a8ce062e12d5b4aafdf91b108