cleanup life cycle issue of PlaybackGlue in PlaybackFragment.
onDetachedFromHost should be called in onDestroy. A PlaybackGlue
should survive onDestroyView for the case setRetainInstance(true).
When configuration change, we should not need recreate PlaybackGlue.
Add a VideoFragmentTest.

Refactoring Tests to use SingleFragmentTestBase and remove
lots of redundant test Activity classes and layout files.

Changed test mp4 to be h264 Baseline profile so it can play
on API 19.

Bug: 35364233
Test: added VideoFragmentTest VideoSupportFragmentTest

Change-Id: I93bd8cbf449ac0fea4b6b76d10dc611c538283e8