@Decompose
This CL introduces a new annotation @Decompose which can be used
to merge sub fields of a Pojo into the class that includes it.
This is great for 1-1 relationships or more logical grouping.

This also builds the baseline for Relationships but does not
yet make it very easy (e.g. no easy way to handle join fields
which will have the same name in 2 entities).

This CL also fixes a bug in Pojo processor where it was ignoring
super class's private fields. A ticket is created for it to
properly detect errors w/ inherited fields from super:
b/35568142

Bug: 34258173
Test: PojoProcessorTest.kt, DecompositionTest.java
Change-Id: Idd2617b5fd0ab0ac70750237076158cfc166c53e