Index annotation
This CL adds an Index field to the Entity annotation to create
indices on the table. While Index is good for composite indices,
it might be too verbose for indexing individual fields. For those,
we add a new field to ColumnInfo#index.

When a @Decomposed field has an index, they are dropped but a
warning is printed for each of them.

When a super class has an index, it is dropped but a warning
is printed for each of them. There is also a new inheritSuperIndices
field in the @Entity annotation to inherit these indices.

We also need to change @PrimaryKey to obey these rules. Created
b/35706557 for that.

I've also fixed a bug in Database Impl generation where we were
generating the wrong class for nested classes.

Bug: 35594281
Test: IndexTest.kt, IndexingTest.java, EntityProcessorTest.kt
Change-Id: I9422f4be4e5d59ed7f2f5f98c18bf3966be0496f