Physics-based animation - SpringAnimation
This CL introduced physics-based animation. Specifically,
SpringAnimation and its base class DynamicAnimation.

This new set of Animation APIs track velocity of the
animation, and they no longer assume a fixed duration.
As a result, interruptions to animations are now handled
much more smoothly, as the velocity change is always
continuous.

SpringAnimation is driven by a SpringForce that calculates
the update of the animation's value and velocity on each
animation pulse until the spring reaches equilibrium.

Test: 100% public API coverage in this CL
BUG: 30999909
Change-Id: Ic47a453d5134ab4092def57a303eb039a3cd95cc