AudioAttributesCompat.
Some APIs accept only AudioAttributes and do not allow
legacy stream types. This class is the first step toward
remedying that; it is a compat lib implementation of the
AudioAttributes structure and type<->stream conversion.

On API21 and later it is a simple wrapper around platform
AudioAttributes. Prior to API21 it holds a stream type.

Future compat versions of other APIs should accept
AudioAttributesCompat where the framework version accepts
AudioAttributes.

Test: ./gradlew support-media-compat:connectedCheck --info --daemon \
	-Pandroid.testInstrumentationRunnerArguments.class=android.support.v4.media.AudioAttributesCompatTest
Change-Id: Ic1b22412bf27f662f7a0a6a97a81d7b2036e5541