Add missing UnsupportedOperationException for queue management
When calling queue management APIs, an UnsupportedOperationException
should be thrown when the session does not support queue managements.
However, this logic is missing in API 21 and after. This CL adds the
missing exceptions. Also, this CL adds more information in the JavaDoc
how developers can check whether the session supports queue management
methods.

Bug: 35952051
Test: Run MediaControllerCompatTest in API 19-25
Change-Id: I6a0753382bd77e6f6ee37eac1a230b2aa389cad3