leanback: fix initial selection of playback button
Because initially when Media is not ready we skipped
updatePlaybackState(), we have initially two buttons
thumbs up and thumbs down in the adapter.
The UI makes initially selection to #2.

And later play/ rewind / fastforward button is added when
hasValidMedia() becomes true, selection is still on
button #2.

This CL calls updateControlsButton during the setup
and add all play buttons even when media is invalid.

We no longer allow onStateChanged() to add/remove play related
actions, afaik no app is relying on this behavior.

Bug: 36071311
Test: PlaybackContrlGlueTest#testWithoutValidMedia

Change-Id: I36d9404ff8ee61d10d10f77ca6b9fdab3aa69d07