Fully support embedded navigation graphs
Adds support for <include app:graph="@xml/other_graph" />
in navigation graph XMl files. Finding nodes and actions now
recursively tranverse up the graph hierarchy looking for the
given node/action (the one restriction is that the
startDestination of a NavGraph must still be a direct child
of the NavGraph).

This required adding an explicit getParent() method to
NavDestination. To ensure that getParent().findNode(getId()) == this,
NavGraph now enforces a null parent when calling addDestination.
The remove() method now nulls out the parent.

Test: ./gradlew assemble
Change-Id: Ide8b33d6f4aad5b2a3a920f8259d20579a3b5ebb