MediaRouter: Do not use remote volume handling for BT routes
Although a BT route is not a remote route, MediaRouter calls
MediaSessionCompat.setPlaybackToRemote() whenever a BT route is 
selected. As a result, the volume dialog in system UI shows two volume 
sliders for the one BT route, which is weird. This CL fixes the problem 
by not using remote volume handling for BT routes.

Bug: 34919351
Test: Checked the volume UI with Support7Demos in API 19-25 and 
      Android O.
Change-Id: Ie92e75ba6625c770814adf051592e58323b60466