leanback: fix item listener for preteched row
leanback had assumption that items on a row was bound after row is
attached to window, row gets OnItemViewClickedListener when it
is attached to window. This assumption is broken when pretch a row.
The first few items of that row is bound before row is attached to
window at that time there is no OnItemViewClickedListener on the row.

Bug: 36665531
Test: RowsFragmentTest#prefetchChildItemsBeforeAttach
Change-Id: I8a88ad955773b533ef362b6e759ce4d88aa88037