Don’t update version and modified fields in font
Keeping the font version as it is, since it provides a link between
compat font and the original font.

Changing the modified field in the font causes the font ttf file to
change, which makes it harder to understand if something really changed
in the font. Since we keep the font version same, we can keep the last
modified field the same.

Test: gradle support-emoji:connectedAndroidTest
      ./emoji/core/scripts/createfont.py with font and unicode files

Change-Id: I9ad296854b5fbe64b1a0e5ca4ae2795bfef53725