Navigation: Support popping multiple destinations
Add NavOptions to pop up to a specific destination,
potentially removing multiple destinations from the back
stack through a single navigation action.

Also includes a boolean to make the popUpTo operation
inclusive or not. This allows us to support the same
level of flexibility as FragmentManager provides.

To get the correct behavior for NavHelper's global
navigation actions, added basic support for singleTop
in FragmentNavigator which simply replaces the fragment
without adding another instance of it to the back stack
if the fragment to add was already on the top of the stack.

Test: testapp global navigation clears back stack correctly

BUG: 36818083
Change-Id: I232f864618af8682e51a8847a57c9fb8759e3ec6