leanback: fix crash late assigning Adapter that has stable ids
We cannot assign a ItemBridgeAdapter to RecyclerView
and later RowsFragment.setAdapter() to change hasStableId
to true. RecylcerView will crash.

So BaseRowsFragment have to delay the
RecyclerView.setAdapter(itemBridgeAdapter)
until app pass in a adapter.

However, this has a side effect on FocusHighlightHelper.
setupHeaderItemFocusHighlight(VerticalGridView).  We no longer
guarantee the VerticalGridView will have an ItemBridgeAdapter
assigned to it in onViewCreated().  So we have to introduce
new APIs to expose getBridgeAdapter() and let FocusHighlightHelper
use ItemBridgeAdapter.

Bug: 37515649
Test: HeadersFragmentTest#disableScaleInConstructor
RowsFragmentTest#changeHasStableIdToTrueAfterViewCreated

Change-Id: I1137b81ced68a9d9a8fb13ca689882b9a39d09c7