New JobIntentService.
This is a new, modern replacement for IntentService.  They key feature
is it uses JobScheduler when running on O+ platforms.  You only need
to implement one service class, which will run as a job on O+ or a plain
service on pre O.  This also removes the use of WakefulBroadcastReceiver,
as it builds wake lock handling into itself (since starting with O, when
using JobScheduler, the app no longer needs to take care of this).

Test: added test, runs on O and N.
Change-Id: I08062bc568d563c70eb49154bff80788865354b2