Fix missing italics in AppCompatTextView
The issue was that we only read the textStyle attribute when
there was the fontFamily attribute present. Some users might
only set the textStyle, and rely on the default fontFamily
set via the default TextAppearance set on the TextView.

This regression was introduced when we started handling all
Typeface assignments via AppCompat instead of relying on the
framework.

Also fixed an adjacent issue where we were not loading String
typeface definitions on a restricted context. Loading system
fonts via String is fine in this case, just not resources fonts.

Bug: 38427057
Test: AppCompatTextViewTest#testTextStyle_setTextStyleInStyle
Change-Id: I470eb9e9d9b18aa3caa09dac12433dcc5a4a2eba