Content paging compat + Load cursors in background.
Adds support for loading cursors in background.
Synthesize paged cursors from unpaged results when paging is requested.
Try to be more efficient than MatrixCursor: InMemoryCursor, avoids a lot of auto-boxing associated with building MatrixCursor.
Adds a very basic demo app that pages through content (but with buttons, not in response to scrolling).

Test: Added tests for new classes. They're awesome!
Change-Id: I024ee0d4a79b87816efd6ad8091bf6de7daf6995