Replace fonts of unknown origin
There are fonts that we are not sure where they came from,
replace with fonts we created ourselves.

The res/font/ folder does not build with gradle when there is
a ttx file in it, so fonts there have been copied from elsewhere
in the tree and contain copyright and licensing inside the font.
There is also a new readme in the root folder for the test projects
using these to point this out.

Also fix a small issue where support tests can't be run on an O
device that ran coretests as the provider authorities colide.
Renamed the support authority to fix this.

Bug: 37726227
Test: ran all support-compat and support-appcompat-v7 tests
Change-Id: I7948fcbd3dbe8b267fc1b51a24a128aae083d213