Add toString dumps to most RV exceptions
Fixes: 38339683
Test: RecyclerviewBasicTest

Add Adapter and LLM toString to RV's toString, and print this with most
user-visible exceptions. This should make it easier to figure out which
RecyclerView is throwing an exception in production, instead of having to
guess based on stack trace depth.

This only covers exceptions in RecyclerView.java - can consider expanding to
other exceptions as necessary.

Change-Id: I94fd16039bb9113c0aa66cbd22ac7f2be6fb564d