Consider nullable types for known column types
When we can get the column type from sqlite at compile time,
we were only considering primitives which is a problem since
the column might be nullable.

This CL fixes that issue by adding nullables into the list.

I've also removed the conveters from basic type to/from string
since it was only useful when we were using the old sqlite API
that requires parameters to be strings. Right now, it just
serves as a misleading edge between type nodes.

Bug: 62100716
Test: TypeAdapterStoreTest, SimpleReadWriteTest
Change-Id: Ifb9c9d41aa694d7a3e1a04b617bbe3dc6e473839