Introduces more general smooth scrolling API for RecyclerView snap helpers.
This change adds a createScroller() method to SnapHelper which wraps the existing createSnapScroller() method, but widens the return type to SmoothScroller instead of LinearSmoothScroller.  This allows subclasses more customization of the scrolling behavior; the only usage of this method in SnapHelper was assigning the returned value to a variable with the wider type anyway so there is no behavior change.  The existing method is marked as deprecated.

Change-Id: Idc6028d2a92ef650d1dba329fce665ce7efa544b