Fixed keyboard autoclosing and wrong position for EditTexts
RV focused on the wrong item when keyboard popped up/or auto-closed for
an RV with a non-exact measure since RV shrank and the anchor focused
child was not updated on the 2nd measure pass and hence not laid out
since availableSpace in LayoutState became negative for this out-of-bound
anchor.
Updated LLM to ensure that anchor child is always laid out if it's the
focused view when it falls outside the viewport on a layout pass.
This issue is now fixed for both LLM and GLM.

Bug: 36082993
Bug: 62373324
Bug: 34887461
Test: ./gradlew support-recyclerview-v7:connectedCheck
-Pandroid.testInstrumentationRunnerArguments.class=android.support.v7.widget.LinearLayoutManagerTest#editTextVisibility
./gradlew support-recyclerview-v7:connectedCheck
-Pandroid.testInstrumentationRunnerArguments.class=android.support.v7.widget.GridLayoutManagerTest#editTextVisibility
Change-Id: I4350d6b2aa4c67fdbc58cdc201786992b6943564