AutoSize Button - support library impl
Introduces functionality to auto-size text on a Button
in AppCompat.

In the framework the classes that support AutoSize are
the ones that extend the base TextView class. Button is
one of them and gets the auto-size functionality for free,
however in AppCompat, due to the chain of inheritance the
AppCompatButton does not inherit the autoSize behavior
because AppCompatButton extends Button (and not
AppCompatTextView).

Isolated the current non-textview-centric
AppCompatTextViewAutoSizeTests in AppCompatBaseAutoSizeTest
and reused them in AppCompatButtonAutoSizeTest.

Note that at this time there is no ButtonCompat thus
we do not have public APIs for configuring auto-size
on a support button (but only @hidden in AppCompatButton),
so can only do XML configuration.

Bug: 63108956
Test: ./gradlew support-appcompat-v7:connectedCheck --info --daemon
-Pandroid.testInstrumentationRunnerArguments.class=android.support.v7.widget.AppCompatButtonAutoSizeTest

Change-Id: I870bf025f3b905011ec992efa107bca17ae9b39f